api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
mysql-connector-python==9.0.0
redis==5.0.8
